backend/requirements/common.txt
Django==4.0.1
django-environ==0.8.1
django-filter==21.1
djoser==2.1.0
drf-extra-fields==3.2.1
djangorestframework==3.13.1
psycopg2-binary==2.8.6
Pillow==9.0.0
gunicorn==20.0.4